requirements.txt
Django==1.8.4
djangorestframework==3.2.3
djangorestframework-csv==2.1.0
Pillow==3.0.0
ipython==5.5.0
pandas==0.23.4
numpy==1.15.2
requests==2.8.1
threadpool==1.3.2
rarfile==3.0
aircv==1.4.6
PyMuPDF==1.14.4
opencv-python==4.1.0.25